requirements.txt
beautifulsoup4>=4.12
openai>=1.40
httpx>=0.27
python-dotenv>=1.0
psycopg[binary]>=3.1
